move useQueryWithPermissions into its own file
